refactor: move general utils
