+++ x/ecocredit/server/core/create_batch.go
@@ -2,6 +2,7 @@ package core
 
 import (
 	"context"
+	"github.com/regen-network/regen-ledger/x/ecocredit/server"
 
 	"google.golang.org/protobuf/types/known/timestamppb"
 
@@ -57,7 +58,7 @@ func (k Keeper) CreateBatch(ctx context.Context, req *core.MsgCreateBatch) (*cor
 		return nil, err
 	}
 
-	creditType, err := GetCreditTypeFromBatchDenom(ctx, k.stateStore, k.params, batchDenom)
+	creditType, err := server.GetCreditTypeFromBatchDenom(ctx, k.stateStore, k.params, batchDenom)
 	if err != nil {
 		return nil, err
 	}
@@ -65,7 +66,7 @@ func (k Keeper) CreateBatch(ctx context.Context, req *core.MsgCreateBatch) (*cor
 
 	tradableSupply, retiredSupply := math.NewDecFromInt64(0), math.NewDecFromInt64(0)
 	for _, issuance := range req.Issuance {
-		decs, err := getNonNegativeFixedDecs(maxDecimalPlaces, issuance.TradableAmount, issuance.RetiredAmount)
+		decs, err := server.GetNonNegativeFixedDecs(maxDecimalPlaces, issuance.TradableAmount, issuance.RetiredAmount)
 		if err != nil {
 			return nil, err
 		}
