+++ x/ecocredit/server/core/utils.go
@@ -5,10 +5,6 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	ecocreditv1 "github.com/regen-network/regen-ledger/api/regen/ecocredit/v1"
-
-	"github.com/regen-network/regen-ledger/types/math"
-	"github.com/regen-network/regen-ledger/x/ecocredit"
 )
 
 // assertClassIssuer makes sure that the issuer is part of issuers of given classID.
@@ -26,44 +22,4 @@ func (k Keeper) assertClassIssuer(goCtx context.Context, classID uint64, issuer 
 		return sdkerrors.ErrUnauthorized.Wrapf("%s is not an issuer for the class", issuer)
 	}
 	return nil
-}
-
-// getCreditType searches for a credit type that matches the given abbreviation within a credit type slice.
-func getCreditType(ctAbbrev string, creditTypes []*ecocredit.CreditType) (ecocredit.CreditType, error) {
-	//creditTypeName = ecocredit.NormalizeCreditTypeName(creditTypeName)
-	for _, creditType := range creditTypes {
-		// credit type name's stored via params have enforcement on normalization, so we can be sure they will already
-		// be normalized here.
-		if creditType.Abbreviation == ctAbbrev {
-			return *creditType, nil
-		}
-	}
-	return ecocredit.CreditType{}, sdkerrors.ErrInvalidType.Wrapf("%s is not a valid credit type", ctAbbrev)
-}
-
-// GetCreditTypeFromBatchDenom extracts the classId from a batch denom string, then retrieves it from the params.
-func GetCreditTypeFromBatchDenom(ctx context.Context, store ecocreditv1.StateStore, k ParamKeeper, denom string) (ecocredit.CreditType, error) {
-	sdkCtx := sdk.UnwrapSDKContext(ctx)
-	classId := ecocredit.GetClassIdFromBatchDenom(denom)
-	classInfo, err := store.ClassInfoStore().GetByName(ctx, classId)
-	if err != nil {
-		return ecocredit.CreditType{}, err
-	}
-	p := &ecocredit.Params{}
-	k.GetParamSet(sdkCtx, p)
-	return getCreditType(classInfo.CreditType, p.CreditTypes)
-}
-
-// getNonNegativeFixedDecs takes an arbitrary amount of decimal strings, and returns their corresponding fixed decimals
-// in a slice.
-func getNonNegativeFixedDecs(precision uint32, decimals ...string) ([]math.Dec, error) {
-	decs := make([]math.Dec, len(decimals))
-	for i, decimal := range decimals {
-		dec, err := math.NewNonNegativeFixedDecFromString(decimal, precision)
-		if err != nil {
-			return nil, err
-		}
-		decs[i] = dec
-	}
-	return decs, nil
 }
