+++ x/ecocredit/server/marketplace/keeper.go
@@ -1,11 +1,10 @@
 package marketplace
 
 import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/params/types"
 	marketplacev1 "github.com/regen-network/regen-ledger/api/regen/ecocredit/marketplace/v1"
 	ecocreditv1 "github.com/regen-network/regen-ledger/api/regen/ecocredit/v1"
 	"github.com/regen-network/regen-ledger/x/ecocredit"
+	"github.com/regen-network/regen-ledger/x/ecocredit/server"
 )
 
 // TODO: Revisit this once we have proper gas fee framework.
@@ -16,14 +15,10 @@ type Keeper struct {
 	stateStore marketplacev1.StateStore
 	coreStore  ecocreditv1.StateStore // TODO: do we want a ref to the core's store? or should we pass it a keeper iface?
 	bankKeeper ecocredit.BankKeeper
-	params     ParamKeeper
+	params     server.ParamKeeper
 }
 
-type ParamKeeper interface {
-	GetParamSet(ctx sdk.Context, ps types.ParamSet)
-}
-
-func NewKeeper(ss marketplacev1.StateStore, cs ecocreditv1.StateStore, bk ecocredit.BankKeeper, params ParamKeeper) Keeper {
+func NewKeeper(ss marketplacev1.StateStore, cs ecocreditv1.StateStore, bk ecocredit.BankKeeper, params server.ParamKeeper) Keeper {
 	return Keeper{
 		coreStore:  cs,
 		stateStore: ss,
